+++ src/shared/Network/EPOLL/SocketMgrLinux.cpp
@@ -6,7 +6,7 @@
  *
  */
 
-#include "Network.h"
+#include "../Network.h"
 
 #ifdef CONFIG_USE_EPOLL
