+++ src/shared/Network/IOCP/SocketWin32.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "CommonTypes.hpp"
-#include "Network.h"
+#include "../Network.h"
 
 #ifdef CONFIG_USE_IOCP
