+++ src/shared/Network/KQUEUE/SocketMgrFreeBSD.cpp
@@ -6,7 +6,7 @@
  *
  */
 
-#include "Network.h"
+#include "../Network.h"
 
 #ifdef CONFIG_USE_KQUEUE
