+++ src/shared/Network/KQUEUE/SocketOpsFreeBSD.cpp
@@ -6,9 +6,9 @@
  *
  */
 
-#include "Network.h"
+#include "../Network.h"
 
-#ifdef CONFIG_USE_EPOLL
+#ifdef CONFIG_USE_KQUEUE
 
 namespace SocketOps
 {
@@ -59,16 +59,6 @@ namespace SocketOps
         return (setsockopt(fd, SOL_SOCKET, SO_RCVBUF, (const char*)&size, sizeof(size)) == 0);
     }
 
-    // Set internal timeout.
-    bool SetTimeout(SOCKET fd, uint32 timeout)
-    {
-        struct timeval to;
-        to.tv_sec = timeout;
-        to.tv_usec = 0;
-        if(setsockopt(fd, SOL_SOCKET, SO_SNDTIMEO, (const char*)&to, (socklen_t)sizeof(to)) != 0) return false;
-        return (setsockopt(fd, SOL_SOCKET, SO_RCVTIMEO, (const char*)&to, (socklen_t)sizeof(to)) == 0);
-    }
-
     // Closes a socket fully.
     void CloseSocket(SOCKET fd)
     {
@@ -82,6 +72,17 @@ namespace SocketOps
         uint32 option = 1;
         if (setsockopt(fd, SOL_SOCKET, SO_REUSEADDR, (const char*)&option, 4) < 0)
             printf("SO_REUSEADDR setsockopt error\n");
+    }
+
+    // Set internal timeout.
+    bool SetTimeout(SOCKET fd, uint32 timeout)
+    {
+        struct timeval to;
+        to.tv_sec = timeout;
+        to.tv_usec = 0;
+        if(setsockopt(fd, SOL_SOCKET, SO_SNDTIMEO, (const char*)&to, (socklen_t)sizeof(to)) != 0)
+            return false;
+        return (setsockopt(fd, SOL_SOCKET, SO_RCVTIMEO, (const char*)&to, (socklen_t)sizeof(to)) == 0);
     }
 }
